The Incredible Adventure! 2015
Basically a clone of Super Mario I did for a game design class. Worked
in Unity and coded in C#.
I was going to make custom sprites but ended up not having time, so some
temp sprites were found via google. There are a lot of things I wanted
to add to this but didn't have time to do. Perhaps I'll continue to work
on it in the future just for fun.